requirements.txt
pyTelegramBotAPI~=4.27.0
requests~=2.32.4
psycopg2-binary~=2.9.10
numpy~=2.1.3
tensorflow~=2.19.0
matplotlib~=3.10.3
keras~=3.10.0
tqdm~=4.67.1
fastapi~=0.116.0
pillow~=11.2.1